+++ CHANGELOG.md
@@ -15,6 +15,8 @@
     * Mailer refactoring - move to `Costlocker\Reports\Export`, catch `Swift_Mailer` exceptions, don't unlink sent file
     * **Google Drive** - upload report to Drive folder, load `personsSettings` from Drive
     * Pretty XLS filename (without `(`, `)` and multiple dashes `--`)
+* _Bugfixes_
+    * Fix exporting XLS with more than 26 columns
 
 ## v1.0.1 (_2017-04-10_)